refactor: extract manual ahjo index to it's own component
